+++ src/pages/_app.tsx
@@ -1,5 +1,5 @@
 import type { AppProps } from 'next/app';
-import '../src/index.css';
+import '../index.css';
 import { DndProvider } from 'react-dnd';
 import { HTML5Backend } from 'react-dnd-html5-backend';
