+++ src/pages/index.tsx
@@ -1,6 +1,6 @@
 import dynamic from 'next/dynamic';
 
-const NextApp = dynamic(() => import('../src/NextApp'), { ssr: false });
+const NextApp = dynamic(() => import('../NextApp'), { ssr: false });
 
 export default function Home() {
   return <NextApp />;
